+++ vlm/backends/cpu/include/vlm_backend_cpu.hpp
@@ -22,9 +22,9 @@ class BackendCPU : public Backend {
         void compute_rhs(const FlowData& flow, const std::vector<f32>& section_alphas) override; 
         void lu_factor() override;
         void lu_solve() override;
-        f32 compute_coefficient_cl(const FlowData& flow, const f32 area, const u32 j, const u32 n) override;
-        linalg::alias::float3 compute_coefficient_cm(const FlowData& flow, const f32 area, const f32 chord, const u32 j, const u32 n) override;
-        f32 compute_coefficient_cd(const FlowData& flow, const f32 area, const u32 j, const u32 n) override;
+        f32 compute_coefficient_cl(const FlowData& flow, const f32 area, const u64 j, const u64 n) override;
+        linalg::alias::float3 compute_coefficient_cm(const FlowData& flow, const f32 area, const f32 chord, const u64 j, const u64 n) override;
+        f32 compute_coefficient_cd(const FlowData& flow, const f32 area, const u64 j, const u64 n) override;
         void compute_delta_gamma() override;
 };
