+++ vlm/include/vlm_mesh.hpp
@@ -17,10 +17,10 @@ class Mesh {
     // (stored in span major order)
     SoA_3D_t<f32> v; // size (ncw+1)*(ns+1)
     // Offsets for indexing in connectivity array for each panel
-    std::vector<u32> offsets = {}; // size nc*ns + 1
+    std::vector<u64> offsets = {}; // size nc*ns + 1
     // Panel-vertex connectivity
     // vertices indices of a panel: connectivity[offsets[i]] to connectivity[offsets[i+1]]
-    std::vector<u32> connectivity = {}; // size 4*(nc*ns)
+    std::vector<u64> connectivity = {}; // size 4*(nc*ns)
 
     // Panels metrics (stored span major order)
     // ---------------------
@@ -33,9 +33,9 @@ class Mesh {
 
     // Structured dimensions
     // ---------------------
-    u32 nc = 1; // number of panels chordwise
-    u32 ns = 1; // number of panels spanwise
-    const u32 nw = 1; // number of panels in chordwise wake
+    u64 nc = 1; // number of panels chordwise
+    u64 ns = 1; // number of panels spanwise
+    const u64 nw = 1; // number of panels in chordwise wake
 
     // Analytical quanities when provided
     // ---------------------
@@ -48,23 +48,23 @@ class Mesh {
     void compute_connectivity(); // fill offsets, connectivity
     void compute_metrics_wing(); // fill colloc, normal, area
     void compute_metrics_wake();
-    void compute_metrics_i(u32 i);
-    u32 nb_panels_wing() const;
-    u32 nb_panels_total() const;
-    u32 nb_vertices_wing() const;
-    u32 nb_vertices_total() const;
-    f32 panels_area(const u32 i, const u32 j, const u32 m, const u32 n) const;
-    f32 panels_area_xy(const u32 i, const u32 j, const u32 m, const u32 n) const;
-    f32 panel_width_y(const u32 i, const u32 j) const;
-    f32 strip_width(const u32 j) const;
-    f32 chord_length(const u32 j) const;
-    f32 chord_mean(const u32 j, const u32 n) const;
+    void compute_metrics_i(u64 i);
+    u64 nb_panels_wing() const;
+    u64 nb_panels_total() const;
+    u64 nb_vertices_wing() const;
+    u64 nb_vertices_total() const;
+    f32 panels_area(const u64 i, const u64 j, const u64 m, const u64 n) const;
+    f32 panels_area_xy(const u64 i, const u64 j, const u64 m, const u64 n) const;
+    f32 panel_width_y(const u64 i, const u64 j) const;
+    f32 strip_width(const u64 j) const;
+    f32 chord_length(const u64 j) const;
+    f32 chord_mean(const u64 j, const u64 n) const;
 
     // i panel vertices coordinates
-    linalg::alias::float3 get_v0(u32 i) const; // upper left
-    linalg::alias::float3 get_v1(u32 i) const; // upper right
-    linalg::alias::float3 get_v2(u32 i) const; // lower right
-    linalg::alias::float3 get_v3(u32 i) const; // lower left
+    linalg::alias::float3 get_v0(u64 i) const; // upper left
+    linalg::alias::float3 get_v1(u64 i) const; // upper right
+    linalg::alias::float3 get_v2(u64 i) const; // lower right
+    linalg::alias::float3 get_v3(u64 i) const; // lower left
 
     void io_read(const std::string& filename);
     Mesh(const tiny::Config& cfg);
